# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Cerberus - Remote Server Management",
        "title": "Cerberus API",
        "contact": {},
        "version": "2.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/docker/container-details": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves information about Docker containers from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docker"
                ],
                "summary": "Get Docker container information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Docker container information retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docker.Container"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticates user against an SSH server and returns a JWT token for subsequent API requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Login to SSH and generate JWT token",
                "parameters": [
                    {
                        "description": "SSH login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in and token generated",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to connect to SSH server or generate token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/server-details": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves basic server information like hostname, OS, kernel version, etc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Get basic server details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/server.ServerDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/server-details/cpu-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves detailed CPU information from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Get CPU information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CPU information retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.CPUInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/server-details/disk-usage": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves disk usage information from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Get disk usage information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disk usage information retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.DiskUsage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/server-details/running-processes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves information about running processes on the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Get running processes information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Running processes information retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.ProcessInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "ip",
                "password",
                "port",
                "username"
            ],
            "properties": {
                "ip": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "docker.Container": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "container_id": {
                    "type": "string"
                },
                "created_on": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "names": {
                    "type": "string"
                },
                "ports": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "server.CPUInfo": {
            "type": "object",
            "properties": {
                "address_sizes": {
                    "type": "string"
                },
                "apicid": {
                    "type": "string"
                },
                "bogomips": {
                    "type": "string"
                },
                "cache_alignment": {
                    "type": "string"
                },
                "cache_size": {
                    "type": "string"
                },
                "clflush_size": {
                    "type": "string"
                },
                "core_id": {
                    "type": "string"
                },
                "cpu_cores": {
                    "type": "string"
                },
                "cpu_family": {
                    "type": "string"
                },
                "cpu_mhz": {
                    "type": "string"
                },
                "cpuid_level": {
                    "type": "string"
                },
                "flags": {
                    "type": "string"
                },
                "fpu": {
                    "type": "string"
                },
                "fpu_exception": {
                    "type": "string"
                },
                "initial_apicid": {
                    "type": "string"
                },
                "microcode": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "model_name": {
                    "type": "string"
                },
                "physical_id": {
                    "type": "string"
                },
                "power_management": {
                    "type": "string"
                },
                "processor": {
                    "type": "string"
                },
                "siblings": {
                    "type": "string"
                },
                "stepping": {
                    "type": "string"
                },
                "vendor_id": {
                    "type": "string"
                },
                "wp": {
                    "type": "string"
                }
            }
        },
        "server.DiskUsage": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "string"
                },
                "filesystem": {
                    "type": "string"
                },
                "mounted_on": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "use_percentage": {
                    "type": "string"
                },
                "used": {
                    "type": "string"
                }
            }
        },
        "server.ProcessInfo": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "cpu_consumption": {
                    "type": "string"
                },
                "process_id": {
                    "type": "string"
                },
                "rss": {
                    "type": "string"
                },
                "started": {
                    "type": "string"
                },
                "stat": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "tty": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                },
                "vsz": {
                    "type": "string"
                }
            }
        },
        "server.ServerDetails": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string"
                },
                "kernel_version": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "uptime": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}